Remove vm adapter (#4247)
* remove unused funtion and group imports

* build stdlib-package with ReleaseBundle

* merge vm_adapter to move_vm_ext
