+++ vm/transaction-builder/Cargo.toml
@@ -7,6 +7,7 @@ starcoin-logger = { workspace = true }
 starcoin-vm-types = { workspace = true }
 starcoin-types = { workspace = true }
 stdlib = { workspace = true }
+starcoin-cached-packages = { workspace = true }
 
 [dev-dependencies]
 stest = { workspace = true }
