+++ vm/vm-runtime/src/lib.rs
@@ -10,8 +10,9 @@ pub mod starcoin_vm;
 #[macro_use]
 pub mod counters;
 
-use move_core_types::vm_status::VMStatus;
+use move_core_types::vm_status::{StatusCode, VMStatus};
 pub use move_vm_runtime::{move_vm, session};
+use std::collections::BTreeMap;
 mod access_path_cache;
 mod errors;
 #[cfg(feature = "force-deploy")]
@@ -19,13 +20,12 @@ pub mod force_upgrade_management;
 pub mod move_vm_ext;
 pub mod parallel_executor;
 mod verifier;
-mod vm_adapter;
 
 use crate::metrics::VMMetrics;
+use starcoin_vm_types::block_metadata::BlockMetadata;
+use starcoin_vm_types::transaction::{SignedUserTransaction, TransactionStatus};
+use starcoin_vm_types::write_set::WriteSet;
 use starcoin_vm_types::{
-    access_path::AccessPath,
-    account_address::AccountAddress,
-    language_storage::StructTag,
     state_store::StateView,
     transaction::{Transaction, TransactionOutput},
 };
@@ -45,7 +45,42 @@ pub trait VMExecutor: Send + Sync {
         metrics: Option<VMMetrics>,
     ) -> Result<Vec<TransactionOutput>, VMStatus>;
 }
-/// Get the AccessPath to a resource stored under `address` with type name `tag`
-fn create_access_path(address: AccountAddress, tag: StructTag) -> AccessPath {
-    AccessPath::resource_access_path(address, tag)
+
+#[derive(Debug)]
+pub enum PreprocessedTransaction {
+    UserTransaction(Box<SignedUserTransaction>),
+    BlockMetadata(BlockMetadata),
+}
+
+#[inline]
+pub fn preprocess_transaction(txn: Transaction) -> PreprocessedTransaction {
+    match txn {
+        Transaction::BlockMetadata(b) => PreprocessedTransaction::BlockMetadata(b),
+        Transaction::UserTransaction(txn) => {
+            PreprocessedTransaction::UserTransaction(Box::new(txn))
+        }
+    }
+}
+
+pub(crate) fn discard_error_vm_status(err: VMStatus) -> (VMStatus, TransactionOutput) {
+    let vm_status = err.clone();
+    let error_code = match err.keep_or_discard() {
+        Ok(_) => {
+            debug_assert!(false, "discarding non-discardable error: {:?}", vm_status);
+            vm_status.status_code()
+        }
+        Err(code) => code,
+    };
+    (vm_status, discard_error_output(error_code))
+}
+
+pub(crate) fn discard_error_output(err: StatusCode) -> TransactionOutput {
+    // Since this transaction will be discarded, no writeset will be included.
+    TransactionOutput::new(
+        BTreeMap::new(),
+        WriteSet::default(),
+        vec![],
+        0,
+        TransactionStatus::Discard(err),
+    )
 }
