--- vm/vm-runtime/src/vm_adapter/mod.rs
@@ -1,13 +1,0 @@
-// Copyright (c) The Starcoin Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-mod adapter;
-mod adapter_common;
-
-pub(crate) use {
-    adapter::PublishModuleBundleOption,
-    adapter_common::{
-        discard_error_output, discard_error_vm_status, preprocess_transaction,
-        PreprocessedTransaction,
-    },
-};
